Stubs refactoring
