+++ fsutil/list_stubs.go
@@ -1,3 +1,6 @@
+//go:build !linux || !darwin || !freebsd
+// +build !linux !darwin !freebsd
+
 package fsutil
 
 // ////////////////////////////////////////////////////////////////////////////////// //
