+++ initsystem/initsystem_stubs.go
@@ -1,3 +1,6 @@
+//go:build !linux || !darwin
+// +build !linux !darwin
+
 // Package initsystem provides methods for working with different init systems
 package initsystem
