+++ initsystem/sdnotify/sdnotify_stubs.go
@@ -1,5 +1,5 @@
-//go:build windows || darwin
-// +build windows darwin
+//go:build !linux
+// +build !linux
 
 // Package sdnotify provides methods for sending notifications to systemd
 package sdnotify
