+++ secstr/secstring.go
