+++ secstr/secstring_test.go
