+++ system/sensors/sensors_stubs.go
@@ -1,3 +1,6 @@
+//go:build !linux || !darwin
+// +build !linux !darwin
+
 // Package sensors provide methods for collecting sensors information
 package sensors
