+++ terminal/input/input_stubs.go
@@ -1,3 +1,6 @@
+//go:build !linux || !darwin
+// +build !linux !darwin
+
 // Package input provides methods for reading user input
 package input
